Add details about how to stop the server running
I wondered about how to stop the server when I don't need it anymore. I run the jar by double-clicking on it, but didn't know how to properly stop it. I realized that once you close the browser tab it sends a signal to the server and kills it in about 5 seconds.

I added that detail to the readme, because it wasn't obvious.